Cambios para el Rol Level 2

+++ app/Http/Controllers/Framing/PaymentController.php
@@ -24,11 +24,8 @@ class PaymentController extends Controller
             'type' => ['required'],
         ]);
     }
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
+
     public function index($subcontractor_id, $user_id)
     {
         $subcontractor = Subcontractor::findOrFail($subcontractor_id);
@@ -39,26 +36,17 @@ class PaymentController extends Controller
                 ->get();
         $totalpayments = $payments->sum('amount');
 
-        if (Auth::user()->role != 1){ return redirect('/home'); }
         return view('framing.payments.index')->with(['subcontractor' => $subcontractor, 'payments' => $payments, 'totalpayments' => $totalpayments, 'user_id' => $user_id]); 
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
+
     public function create($subcontractor_id, $user_id)
     {
         return view("framing.payments.create")->with(['subcontractor_id' => $subcontractor_id,'user_id' => $user_id]);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(Request $request, $subcontractor_id, $user_id)
     {
         $this->validator($request->all())->validate();
@@ -86,37 +74,17 @@ class PaymentController extends Controller
         }
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+ 
     public function edit($id, $subcontractor_id, $user_id)
     {
+        if (Auth::user()->role != 1){ return redirect('/home'); }
+
         $payment = Payment::findOrFail($id);
         
         return view("framing.payments.edit")->with(['subcontractor_id' => $subcontractor_id, 'payment' => $payment, 'user_id' => $user_id]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id, $subcontractor_id, $user_id)
     {
         $this->validator($request->all())->validate();
@@ -140,15 +108,11 @@ class PaymentController extends Controller
         }
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy($id, $subcontractor_id, $user_id)
     {
-
+        if (Auth::user()->role != 1){ return redirect('/home'); }
+        
         Payment::destroy($id);
 
         return redirect('/payments/'.$subcontractor_id.'/'.$user_id)->with(['success' => 'Payment successfully delete']);
